refactor(normalize): rename normalizer module to normalizers
